refactor: simplify GPU rendering context and buffer preparation
