+++ sparse_strips/vello_hybrid/examples/gpu.rs
@@ -11,7 +11,7 @@ use peniko::{
     color::palette,
     kurbo::{BezPath, Stroke},
 };
-use vello_hybrid::{GpuRenderContext, GpuSession};
+use vello_hybrid::{GpuSession, RenderContext};
 use wgpu::util::DeviceExt;
 
 use winit::{
@@ -63,9 +63,9 @@ async fn run(event_loop: EventLoop<()>, window: Window) {
 
     let session = GpuSession::new(&device, format);
     // TODO: actually render something
-    let mut render_ctx = GpuRenderContext::new(size.width as usize, size.height as usize);
+    let mut render_ctx = RenderContext::new(size.width as u16, size.height as u16);
     draw_simple_scene(&mut render_ctx);
-    let bufs = render_ctx.harvest2();
+    let bufs = render_ctx.prepare_gpu_buffers();
 
     let config = Config {
         width: size.width,
@@ -161,7 +161,7 @@ async fn run(event_loop: EventLoop<()>, window: Window) {
         .unwrap();
 }
 
-fn draw_simple_scene(ctx: &mut GpuRenderContext) {
+fn draw_simple_scene(ctx: &mut RenderContext) {
     let mut path = BezPath::new();
     path.move_to((10.0, 10.0));
     path.line_to((180.0, 20.0));
@@ -170,10 +170,10 @@ fn draw_simple_scene(ctx: &mut GpuRenderContext) {
     let piet_path = path.into();
     ctx.set_transform(Affine::scale(5.0));
     ctx.set_paint(palette::css::REBECCA_PURPLE.into());
-    ctx.fill(&piet_path, palette::css::REBECCA_PURPLE.into());
+    ctx.fill_path(&piet_path);
     let stroke = Stroke::new(5.0);
     ctx.set_paint(palette::css::DARK_BLUE.into());
-    ctx.stroke(&piet_path, stroke, palette::css::DARK_BLUE.into());
+    ctx.stroke_path(&piet_path);
 }
 
 fn main() {
