+++ sparse_strips/vello_hybrid/src/render.rs
@@ -3,13 +3,14 @@
 
 //! Basic render operations.
 
+use bytemuck::{Pod, Zeroable};
 use kurbo::{Affine, BezPath, Cap, Join, Rect, Shape, Stroke};
 use peniko::color::palette::css::BLACK;
 use peniko::{BlendMode, Compose, Fill, Mix};
-use vello_common::coarse::Wide;
+use vello_common::coarse::{WIDE_TILE_WIDTH, Wide};
 use vello_common::flatten::Line;
 use vello_common::paint::Paint;
-use vello_common::strip::Strip;
+use vello_common::strip::{STRIP_HEIGHT, Strip};
 use vello_common::tile::Tiles;
 use vello_common::{flatten, strip};
 use vello_cpu::fine::Fine;
@@ -177,5 +178,88 @@ impl RenderContext {
         );
 
         self.wide.generate(&self.strip_buf, fill_rule, paint);
+    }
+}
+
+pub struct GpuRenderBuffers {
+    pub strips: Vec<GpuStrip>,
+    pub alphas: Vec<u32>,
+}
+
+#[repr(C)]
+#[derive(Debug, Clone, Copy, Zeroable, Pod)]
+pub struct GpuStrip {
+    x: u16,
+    y: u16,
+    width: u16,
+    dense_width: u16,
+    col: u32,
+    rgba: u32,
+}
+
+impl RenderContext {
+    pub fn prepare_gpu_buffers(&self) -> GpuRenderBuffers {
+        let mut strips: Vec<GpuStrip> = Vec::new();
+        let width_tiles = (self.width).div_ceil(WIDE_TILE_WIDTH);
+        let height_tiles = (self.height).div_ceil(STRIP_HEIGHT);
+        for y in 0..height_tiles {
+            for x in 0..width_tiles {
+                let tile = &self.wide.tiles[y * width_tiles + x];
+                let tile_x = x * WIDE_TILE_WIDTH;
+                let tile_y = y * STRIP_HEIGHT;
+                let bg = tile.bg.to_rgba8().to_u32();
+                if bg != 0 {
+                    let strip = GpuStrip {
+                        x: tile_x as u16,
+                        y: tile_y as u16,
+                        width: WIDE_TILE_WIDTH as u16,
+                        dense_width: 0,
+                        col: 0,
+                        rgba: bg,
+                    };
+                    strips.push(strip);
+                }
+                for cmd in &tile.cmds {
+                    match cmd {
+                        vello_common::coarse::Cmd::Fill(fill) => {
+                            let color: peniko::color::AlphaColor<peniko::color::Srgb> =
+                                match fill.paint {
+                                    Paint::Solid(color) => color,
+                                    _ => peniko::color::AlphaColor::TRANSPARENT,
+                                };
+                            let strip = GpuStrip {
+                                x: (tile_x as u32 + fill.x) as u16,
+                                y: tile_y as u16,
+                                width: fill.width as u16,
+                                dense_width: 0,
+                                col: 0,
+                                rgba: color.to_rgba8().to_u32(),
+                            };
+                            strips.push(strip);
+                        }
+                        vello_common::coarse::Cmd::AlphaFill(cmd_strip) => {
+                            let color: peniko::color::AlphaColor<peniko::color::Srgb> =
+                                match cmd_strip.paint {
+                                    Paint::Solid(color) => color,
+                                    _ => peniko::color::AlphaColor::TRANSPARENT,
+                                };
+                            let strip = GpuStrip {
+                                x: (tile_x as u32 + cmd_strip.x) as u16,
+                                y: tile_y as u16,
+                                width: cmd_strip.width as u16,
+                                dense_width: cmd_strip.width as u16,
+                                col: cmd_strip.alpha_ix as u32,
+                                rgba: color.to_rgba8().to_u32(),
+                            };
+                            strips.push(strip);
+                        }
+                    }
+                }
+            }
+        }
+        GpuRenderBuffers {
+            strips,
+            alphas: self.alphas.clone(),
+        }
     }
 }
